@@ -1,6 +1,6 @@
 name: syncfusion_flutter_calendar
 description: The Flutter Calendar widget has nine built-in configurable views that provide basic functionalities for scheduling and representing appointments/events efficiently.
-version: 25.2.7
+version: 28.1.33
 homepage: https://github.com/syncfusion/flutter-widgets/tree/master/packages/syncfusion_flutter_calendar
 
 screenshots:
@@ -12,23 +12,22 @@ screenshots:
     path: screenshots/flutter-calendar-load-more.gif
 
 environment:
-  sdk: '>=2.17.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  timezone: 0.9.3
+  timezone: 0.10.0
   syncfusion_flutter_core:
     path: ../syncfusion_flutter_core
-
   syncfusion_flutter_datepicker:
     path: ../syncfusion_flutter_datepicker
+      
+      
 
-  intl: '>=0.18.1 <0.20.0'
+  intl: '>=0.18.1 <0.21.0'
 
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
+
 
 flutter:
   assets:
