@@ -269,7 +269,7 @@ class _CustomCalendarScrollViewState extends State<CustomCalendarScrollView>
   late ValueNotifier<_DragPaintDetails> _dragDetails;
   Offset? _dragDifferenceOffset;
   Timer? _timer;
-  double? _viewPortHeight;
+  late double? _viewPortHeight;
 
   @override
   void initState() {
@@ -5634,8 +5634,8 @@ class _CalendarViewState extends State<_CalendarView>
   late double _timeIntervalHeight;
   final UpdateCalendarStateDetails _updateCalendarStateDetails =
       UpdateCalendarStateDetails();
-  ValueNotifier<SelectionDetails?> _allDaySelectionNotifier =
-      ValueNotifier<SelectionDetails?>(null);
+  ValueNotifier<AllDayPanelSelectionDetails?> _allDaySelectionNotifier =
+      ValueNotifier<AllDayPanelSelectionDetails?>(null);
   late ValueNotifier<Offset?> _viewHeaderNotifier;
   final ValueNotifier<Offset?> _calendarCellNotifier =
           ValueNotifier<Offset?>(null),
@@ -5803,7 +5803,8 @@ class _CalendarViewState extends State<_CalendarView>
     /// select the same month cell and move to day view then the view show
     /// calendar cell selection and all day panel selection.
     if (oldWidget.view != widget.view) {
-      _allDaySelectionNotifier = ValueNotifier<SelectionDetails?>(null);
+      _allDaySelectionNotifier =
+          ValueNotifier<AllDayPanelSelectionDetails?>(null);
       final DateTime today = DateTime.now();
       _currentTimeNotifier = ValueNotifier<int>(
           (today.day * 24 * 60) + (today.hour * 60) + today.minute);
@@ -6174,14 +6175,16 @@ class _CalendarViewState extends State<_CalendarView>
       final double scrollPosition = _getScrollPositionForCurrentDate(
           _updateCalendarStateDetails.currentDate!);
       if (scrollPosition == -1 ||
-          _scrollController!.position.pixels == scrollPosition) {
+          (_scrollController != null &&
+              _scrollController!.position.pixels == scrollPosition)) {
         return;
       }
-
-      _scrollController!.jumpTo(
-          _scrollController!.position.maxScrollExtent > scrollPosition
-              ? scrollPosition
-              : _scrollController!.position.maxScrollExtent);
+      if (_scrollController != null) {
+        _scrollController!.jumpTo(
+            _scrollController!.position.maxScrollExtent > scrollPosition
+                ? scrollPosition
+                : _scrollController!.position.maxScrollExtent);
+      }
     });
   }
 
@@ -6214,7 +6217,7 @@ class _CalendarViewState extends State<_CalendarView>
       }
     }
 
-    if (_scrollController!.hasClients) {
+    if (_scrollController != null && _scrollController!.hasClients) {
       if (timeToPosition > _scrollController!.position.maxScrollExtent) {
         timeToPosition = _scrollController!.position.maxScrollExtent;
       } else if (timeToPosition < _scrollController!.position.minScrollExtent) {
@@ -6477,7 +6480,7 @@ class _CalendarViewState extends State<_CalendarView>
     final Widget shadowView = Divider(
       height: 1,
       thickness: 1,
-      color: borderColor.withOpacity(borderColor.opacity * 0.5),
+      color: borderColor.withValues(alpha: borderColor.a * 0.5),
     );
 
     final double timeLabelWidth = CalendarViewHelper.getTimeLabelWidth(
@@ -8744,7 +8747,9 @@ class _CalendarViewState extends State<_CalendarView>
             padding: EdgeInsets.zero,
             controller: _timelineRulerController,
             scrollDirection: Axis.horizontal,
-            physics: const _CustomNeverScrollableScrollPhysics(),
+            physics: widget.isMobilePlatform
+                ? const _CustomNeverScrollableScrollPhysics()
+                : const ClampingScrollPhysics(),
             children: <Widget>[
               RepaintBoundary(
                   child: CustomPaint(
@@ -8775,7 +8780,9 @@ class _CalendarViewState extends State<_CalendarView>
                 padding: EdgeInsets.zero,
                 controller: _scrollController,
                 scrollDirection: Axis.horizontal,
-                physics: const _CustomNeverScrollableScrollPhysics(),
+                physics: widget.isMobilePlatform
+                    ? const _CustomNeverScrollableScrollPhysics()
+                    : const ClampingScrollPhysics(),
                 children: <Widget>[
                   SizedBox(
                       width: width,
@@ -9395,7 +9402,7 @@ class _CalendarViewState extends State<_CalendarView>
       return;
     }
 
-    _allDaySelectionNotifier.value = SelectionDetails(view, date);
+    _allDaySelectionNotifier.value = AllDayPanelSelectionDetails(view, date);
   }
 
   //// Handles the onTap callback for day view cells, all day panel, and view
@@ -11635,13 +11642,13 @@ class _ViewHeaderViewPainter extends CustomPainter {
       if (!isDateWithInDateRange(minDate, maxDate, currentDate)) {
         dayTextStyle = dayTextStyle.copyWith(
             color: dayTextStyle.color != null
-                ? dayTextStyle.color!.withOpacity(0.38)
+                ? dayTextStyle.color!.withValues(alpha: 0.38)
                 : themeData.brightness == Brightness.light
                     ? Colors.black26
                     : Colors.white38);
         dateTextStyle = dateTextStyle.copyWith(
             color: dateTextStyle.color != null
-                ? dateTextStyle.color!.withOpacity(0.38)
+                ? dateTextStyle.color!.withValues(alpha: 0.38)
                 : themeData.brightness == Brightness.light
                     ? Colors.black26
                     : Colors.white38);
@@ -11783,7 +11790,7 @@ class _ViewHeaderViewPainter extends CustomPainter {
           hoveringColor: (themeData.brightness == Brightness.dark
                   ? Colors.white
                   : Colors.black87)
-              .withOpacity(0.04));
+              .withValues(alpha: 0.04));
     }
   }
 
@@ -11800,11 +11807,11 @@ class _ViewHeaderViewPainter extends CustomPainter {
         xPosition + dateXPosition + _dateTextPainter.width >=
             viewHeaderNotifier.value!.dx) {
       final Color hoveringColor = isToday
-          ? Colors.black.withOpacity(0.12)
+          ? Colors.black.withValues(alpha: 0.12)
           : (themeData.brightness == Brightness.dark
                   ? Colors.white
                   : Colors.black87)
-              .withOpacity(0.04);
+              .withValues(alpha: 0.04);
       _drawTodayCircle(
           canvas,
           xPosition + dateXPosition,
@@ -12869,7 +12876,7 @@ class _CurrentTimeIndicator extends CustomPainter {
     final int viewEndMinutes = (timeSlotViewSettings.endHour * 60).toInt();
     DateTime getLocationDateTime = DateTime.now();
 
-    if (!timeZoneLoaded && timeZone == null) {
+    if (!timeZoneLoaded) {
       return;
     }
     if (totalMinutes < viewStartMinutes || totalMinutes > viewEndMinutes) {
@@ -14098,7 +14105,7 @@ class _DraggingAppointmentRenderObject extends RenderBox
     xPosition = dragDetails.position.value!.dx;
     yPosition = dragDetails.position.value!.dy;
     _shadowPainter.color =
-        dragDetails.appointmentView!.appointment!.color.withOpacity(0.5);
+        dragDetails.appointmentView!.appointment!.color.withValues(alpha: 0.5);
 
     final RRect rect = RRect.fromRectAndRadius(
         Rect.fromLTWH(
