@@ -31,7 +31,8 @@ class TimeSlotWidget extends StatefulWidget {
       this.width,
       this.height,
       this.minDate,
-      this.maxDate);
+      this.maxDate,
+      {super.key});
 
   /// Holds the visible dates collection for current time slot view.
   final List<DateTime> visibleDates;
@@ -753,7 +754,7 @@ class _TimeSlotRenderObject extends CustomCalendarRenderObject {
       rect = Rect.fromLTRB(
           leftPosition, topPosition, rightPosition, bottomPosition);
       _linePainter.style = PaintingStyle.fill;
-      _linePainter.color = Colors.grey.withOpacity(0.2);
+      _linePainter.color = Colors.grey.withValues(alpha: 0.2);
       canvas.drawRect(rect, _linePainter);
     }
   }
@@ -797,7 +798,8 @@ class _TimeSlotRenderObject extends CustomCalendarRenderObject {
         timeIntervalHeight;
     _linePainter.style = PaintingStyle.stroke;
     _linePainter.strokeWidth = strokeWidth;
-    _linePainter.color = calendarTheme.selectionBorderColor!.withOpacity(0.4);
+    _linePainter.color =
+        calendarTheme.selectionBorderColor!.withValues(alpha: 0.4);
     left += isRTL ? 0 : timeLabelWidth;
     double height = timeIntervalHeight;
     if (top == 0) {
@@ -835,7 +837,7 @@ class _TimeSlotRenderObject extends CustomCalendarRenderObject {
     for (int i = 0; i < count; i++) {
       final TimeRegionView view = specialRegionBounds[i];
       final CalendarTimeRegion region = view.region;
-      _linePainter.color = region.color ?? Colors.grey.withOpacity(0.2);
+      _linePainter.color = region.color ?? Colors.grey.withValues(alpha: 0.2);
 
       final TextStyle textStyle = themeData.textTheme.bodyMedium!
           .copyWith(fontSize: 14)
