@@ -39,7 +39,8 @@ class MonthViewWidget extends StatefulWidget {
       this.height,
       this.weekNumberStyle,
       this.isMobilePlatform,
-      this.visibleAppointmentNotifier);
+      this.visibleAppointmentNotifier,
+      {super.key});
 
   /// Defines the row count for the month view.
   final int rowCount;
@@ -930,7 +931,7 @@ class _MonthViewRenderObject extends CustomCalendarRenderObject {
     final TextStyle? blackoutDatesStyle = calendarTheme.blackoutDatesTextStyle;
     final TextStyle disabledTextStyle = currentMonthTextStyle.copyWith(
         color: currentMonthTextStyle.color != null
-            ? currentMonthTextStyle.color!.withOpacity(0.38)
+            ? currentMonthTextStyle.color!.withValues(alpha: 0.38)
             : themeData.brightness == Brightness.light
                 ? Colors.black26
                 : Colors.white38);
@@ -1125,7 +1126,8 @@ class _MonthViewRenderObject extends CustomCalendarRenderObject {
         yPosition + cellHeight >= calendarCellNotifier.value!.dy) {
       _linePainter.style = PaintingStyle.stroke;
       _linePainter.strokeWidth = 2;
-      _linePainter.color = calendarTheme.selectionBorderColor!.withOpacity(0.4);
+      _linePainter.color =
+          calendarTheme.selectionBorderColor!.withValues(alpha: 0.4);
       canvas.drawRect(
           Rect.fromLTWH(
               xPosition == 0 ? xPosition + linePadding : xPosition,
